Remove distinct project resource
Instead of creating a new resource just for projects, I think it is
better to just use blog posts with the `project` tag. For projects with
multiple pages, I can create a `project_name` tag for that project and
use it for all subsequent pages. All project websites should just be in
the `gh-pages` branch of that project anyways.